% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/canvec.R
\name{canvec.cachedir}
\alias{canvec.cachedir}
\title{Get the default cache directory, which is the folder rcanvec.cache
in the current working directory. Modify this behaviour by passing
a \code{cachedir} argument to canvec.download(), canvec.load(), or
canvec.qplot().}
\usage{
canvec.cachedir()
}
\description{
Get the default cache directory, which is the folder rcanvec.cache
in the current working directory. Modify this behaviour by passing
a \code{cachedir} argument to canvec.download(), canvec.load(), or
canvec.qplot().
}

