% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/canvec.R
\name{canvec.defaultoptions}
\alias{canvec.defaultoptions}
\title{Get Default Options For Plotting Layers}
\usage{
canvec.defaultoptions(layerid)
}
\arguments{
\item{layerid}{The layer id as defined in \code{canvec_layers$id}}
}
\value{
a \code{list} object that can be passed to \code{canvec.plot()}
}
\description{
Get Default Options For Plotting Layers
}

