% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/canvec.R
\name{canvec.download}
\alias{canvec.download}
\title{Download and extract canvec or canvec+ data to \code{cachedir}.}
\usage{
canvec.download(..., forcedownload = FALSE, forceextract = FALSE,
  cachedir = NULL)
}
\description{
Download and extract canvec or canvec+ data to \code{cachedir}.
}

