% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/canvec.R
\name{canvec.export}
\alias{canvec.export}
\title{Export layerids for NTS reference(s) to path tofolder, automatically renaming layers
based on their associated layerid.}
\usage{
canvec.export(ntsid, tofolder, layerids = NULL, cachedir = NULL,
  overwrite = TRUE)
}
\description{
Export layerids for NTS reference(s) to path tofolder, automatically renaming layers
based on their associated layerid.
}

