% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canvec.R
\name{canvec.findlayer}
\alias{canvec.findlayer}
\title{Get File Prefix of a CanVec Layer}
\usage{
canvec.findlayer(directory, layerid)
}
\arguments{
\item{directory}{A directory where CanVec shapefiles are located.}

\item{layerid}{A single layer id as listed in \code{canvec_layers$id}}
}
\value{
The file prefix of the layer, or \code{NA} if the layer does not exist
}
\description{
Find directory and file prefix for a layer id (as 
listed in \code{canvec_layers$id}) in the directory specified. 
If the layer is not available, a warning will be issued.
}

