% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/canvec.R
\name{canvec.load}
\alias{canvec.load}
\title{Load layerid for NTS reference(s) that were previously downloaded to cachedir.}
\usage{
canvec.load(ntsid, layerid, cachedir = NULL)
}
\description{
Load layerid for NTS reference(s) that were previously downloaded to cachedir.
}

