% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canvec.R
\name{canvec.loadfromdir}
\alias{canvec.loadfromdir}
\title{Load CanVec Data From Directory}
\usage{
canvec.loadfromdir(directory, layerid)
}
\arguments{
\item{directory}{A directory or directories that contain(s) CanVec or CanVec+ data.}

\item{layerid}{A single layer id as listed in \code{canvec_layers$id}}
}
\value{
A sp::Spatial* object loaded from the given shapefile or a \code{list}
        of Spatial* objects if more than one directory is specified.
}
\description{
Load layerid from a directory or directories that contain(s) CanVec data.
}
\seealso{
canvec.load
}

