% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/canvec.R
\name{canvec.plot}
\alias{canvec.plot}
\title{Plot CanVec Spatial Data}
\usage{
canvec.plot(loaded, options = NULL, crs = NULL, add = NULL)
}
\arguments{
\item{loaded}{A Spatial* object or list of Spatial* objects such as those
generated by \code{canvec.load()}}

\item{options}{A \code{list} object with the graphical options to be applied
to the layers specified}

\item{crs}{A CRS (as generated by \code{sp::CRS()}) in which to project the data.}

\item{add}{TRUE if layer or layers should be added to the current plot, FALSE
if all layers should be plotted on a fresh plot (not reccomended) or NULL for
default behaviour, which will create a new plot for the first layer and add
each subsequent layer}
}
\description{
Plot CanVec Spatial Data
}

