% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/canvec.qplot.R
\name{canvec.qplot}
\alias{canvec.qplot}
\title{Quickly Plot CanVec Data}
\usage{
canvec.qplot(ntsid = NULL, bbox = NULL, waterbody = TRUE, river = TRUE,
  contour = FALSE, building = FALSE, road = FALSE,
  waterbody.col = "lightblue", waterbody.border = "lightblue",
  contour.col = "brown", contour.lwd = 0.2, river.col = "lightblue",
  river.lwd = 1, road.col = "black", road.lwd = 0.5, building.pch = ".",
  building.col = "black", plotdata = TRUE, cachedir = NULL, data = NULL,
  atscale = nts.SCALE50K, ...)
}
\arguments{
\item{ntsid}{One or more NTS References as generated by \code{nts()}}

\item{bbox}{A bounding box describing the desired extent. If no \code{ntsid} is
provided, nts(bbox=bbox) will be invoked to find the appropriate NTS Reference(s)}

\item{waterbody}{TRUE if waterbody (lakes) layer should be drawn}

\item{river}{TRUE if river layer should be drawn}

\item{contour}{TRUE if contour layer should be drawn}

\item{building}{TRUE if building layer should be drawn}

\item{road}{TRUE if road layer should be drawn}

\item{waterbody.col}{Colour to fill waterbody polygons}

\item{waterbody.border}{Colour to outline waterbody polygons}

\item{contour.col}{Colour to draw contour lines}

\item{contour.lwd}{Width of contour lines}

\item{river.col}{Colour to draw rivers}

\item{river.lwd}{Width to draw rivers}

\item{road.col}{Colour to draw roads}

\item{road.lwd}{Width to draw roads}

\item{building.pch}{Symbol to draw for buildings}

\item{building.col}{Colour to draw buildings}

\item{plotdata}{TRUE if plot() should be invoked to draw data. FALSE if just loading data
is desired.}

\item{cachedir}{Pass a specific cache directory in which files have been extracted.
Default value is that returned by \code{canvec.cachedir()}}

\item{data}{A list object that contains the loaded Spatial* data to be plotted.
This should always be an object that was returned by \code{canvec.qplot()}}

\item{atscale}{One of \code{nts.SCALE50K} or \code{nts.SCALE250K}}

\item{...}{A list of graphical parameters passed to the inital call to \code{plot()}.
Note that xlim and ylim will generate an error - instead use the bbox argument.}
}
\value{
A list object that contains the Spatial* data that was plotted
}
\description{
Quickly plot basic CanVec data (lakes, rivers, contours, buildings, and roads) with options
to change the plotting style of each. If data does not exist in the cache
it will be downloaded.
}
\examples{
plotdata <- canvec.qplot(nts("21h1"), contour=TRUE, building=TRUE, road=TRUE)
plotdata <- canvec.qplot(bbox=makebbox(45.1, -64.35, 45.05, -64.4),
                        contour=TRUE, building=TRUE, road=TRUE,
                        data=plotdata)
}

