% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/canvec.R
\name{canvec.url}
\alias{canvec.url}
\title{Generate url from a single NTS reference for canvec or canvec+}
\usage{
canvec.url(ntsid, server = "http://ftp2.cits.rncan.gc.ca/pub")
}
\description{
Generate url from a single NTS reference for canvec or canvec+
}

