% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{canvec_layers}
\alias{canvec_layers}
\title{Canvec Layers}
\format{A data frame with six variables: \code{name}, \code{id},
  \code{theme}, \code{filename}, \code{geometry}, and (\code{geometry_ext}.
  These are used to generate pretty filenames and provide human-readable
  access to canvec and canvec+ data.}
\usage{
canvec_layers
}
\description{
As documented by the Canvec and canvec+ documentation found at
http://ftp2.cits.rncan.gc.ca/pub/canvec+/doc/CanVec+_distribution_formats.pdf
}
\keyword{datasets}

