% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ntstiles.R
\name{nts.bbox}
\alias{nts.bbox}
\title{Get Bounding Box of an NTS Reference}
\usage{
nts.bbox(ntsid)
}
\arguments{
\item{ntsid}{}
}
\description{
Calculates the bounding box in latitude/longitude described by
a particular NTS Reference.
}

