% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntstiles.R
\name{nts.idat}
\alias{nts.idat}
\title{Get NTS Reference At A Location}
\usage{
nts.idat(lat, lon, atscale)
}
\arguments{
\item{lat}{A scalar or vector of latitude values}

\item{lon}{A scalar or vector of longitude values of same length as \code{lat}}

\item{atscale}{One of \code{nts.SCALESERIES}, \code{nts.SCALE250K}, 
or \code{nts.SCALE50K}}
}
\value{
A \code{list} of NTS References
}
\description{
Get NTS Reference(s) based on location at a given scale.
}
\seealso{
\link{nts}
}

