% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/nts.R
\name{nts}
\alias{nts}
\title{Generate one or more NTS references based on arguments provided.}
\usage{
nts(..., lat = NULL, lon = NULL, bbox = NULL, atscale = nts.SCALE50K)
}
\arguments{
\item{...}{An arbitrary number of strings in the form 21h1 to be parsed}

\item{lat}{A vector of latitude values}

\item{lon}{A vector of longitude values}

\item{bbox}{A bounding box in the form returned by sp::bbox()}

\item{atscale}{An integer value describing scale, either 1 (250k) or 2 (50k)}
}
\value{
one or more NTS references.
}
\description{
Generate one or more NTS references based on arguments provided.
}

