% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntstiles.R
\docType{data}
\name{nts.SCALE50K}
\alias{nts.SCALE50K}
\title{A constant denoting NTS Map Sheet (1:50k) scale (2)
 @export}
\format{An object of class \code{numeric} of length 1.}
\usage{
nts.SCALE50K
}
\description{
A constant denoting NTS Map Sheet (1:50k) scale (2)
 @export
}
\keyword{datasets}

