% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ntstiles.R
\docType{data}
\name{nts.SCALESERIES}
\alias{nts.SCALESERIES}
\title{A contstant denoting NTS Series scale (0)}
\format{\preformatted{ num 0
}}
\usage{
nts.SCALESERIES
}
\description{
A contstant denoting NTS Series scale (0)
}
\keyword{datasets}

