% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/nts.R
\name{ntsstring}
\alias{ntsstring}
\title{Generate NTS strings from NTS references or other arguments to generate
such a list from nts() based on arguments provided.}
\usage{
ntsstring(ntsid = NULL, lat = NULL, lon = NULL, bbox = NULL,
  atscale = nts.SCALE50K)
}
\arguments{
\item{lat}{A vector of latitude values}

\item{lon}{A vector of longitude values}

\item{bbox}{A bounding box in the form returned by sp::bbox()}

\item{atscale}{An integer value describing scale, either 1 (250k) or 2 (50k)}

\item{...}{An arbitrary number of NTS references in the form c("021", "H", "01")
to be converted}
}
\value{
one or more NTS references.
}
\description{
Generate NTS strings from NTS references or other arguments to generate
such a list from nts() based on arguments provided.
}

