% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/bboxtools.R
\name{searchbbox}
\alias{searchbbox}
\title{Query Google For A Bounding Box}
\usage{
searchbbox(querystring)
}
\arguments{
\item{querystring}{The search query}
}
\value{
A 2x2 matrix describing a bounding box like that returned by \code{sp::bbox()}
}
\description{
Use the \code{ggmaps::geocode} function to retreive a bounding box for a given query.
}

