% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/canvec.R
\name{canvec.cachedir}
\alias{canvec.cachedir}
\title{Get Cache Directory}
\usage{
canvec.cachedir()
}
\value{
A character string of the cache directory path
}
\description{
Get the default cache directory, which is the folder rcanvec.cache
in the current working directory. Modify this behaviour by passing
a \code{cachedir} argument to \code{canvec.download()}, \code{canvec.load()}, or
\code{canvec.qplot()}.
}

