% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/canvec.qplot.R
\name{canvec.qplot}
\alias{canvec.qplot}
\title{Quickly Plot Canvec Data}
\usage{
canvec.qplot(ntsid = NULL, bbox = NULL, layers = c("waterbody", "contour",
  "river", "building", "road"), options = NULL, data = NULL,
  cachedir = NULL, plotdata = TRUE, atscale = nts.SCALE50K, ...)
}
\arguments{
\item{ntsid}{One or more NTS References as generated by \code{nts()}}

\item{bbox}{A bounding box describing the desired extent. If no \code{ntsid} is
provided, nts(bbox=bbox) will be invoked to find the appropriate NTS Reference(s)}

\item{layers}{A list of layers as defined in \code{canvec_layers$id} in the order in
which they should be plotted}

\item{options}{A list object containing the options for each layer in the form
\code{options$layerid <- list(col="lightblue")}}

\item{data}{A list object that contains the loaded Spatial* data to be plotted.
This should always be an object that was returned by \code{canvec.qplot()}}

\item{cachedir}{Pass a specific cache directory in which files have been extracted.
Default value is that returned by \code{canvec.cachedir()}}

\item{atscale}{One of \code{nts.SCALE50K} (CanVec data) or \code{nts.SCALE250K} (CanVec+ data)}

\item{...}{A list of graphical parameters passed to the inital call to \code{plot()}.}
}
\value{
A list object that contains the Spatial* data that was plotted
}
\description{
Quickly plot CanVec data with options to change the plotting style of each.
If data does not exist in the cache it will be downloaded.
}

