% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/canvec.R
\name{canvec.qplot}
\alias{canvec.qplot}
\title{Quickly Plot CanVec Data}
\usage{
canvec.qplot(ntsid = NULL, bbox = NULL, waterbody = TRUE, river = TRUE,
  contour = FALSE, building = FALSE, road = FALSE,
  waterbody.col = "lightblue", waterbody.border = "lightblue",
  contour.col = "brown", contour.lwd = 0.2, river.col = "lightblue",
  river.lwd = 1, road.col = "black", road.lwd = 0.5, building.pch = ".",
  building.col = "black", plotdata = TRUE, cachedir = NULL, data = NULL,
  atscale = nts.SCALE50K, ...)
}
\description{
Quickly plot one more NTS references or automatically look up references based on a bbox
(in the same format returned by sp::bbox())
}

