% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/canvec.R
\name{canvec.url}
\alias{canvec.url}
\title{Get CanVec or CanVec+ data URL}
\usage{
canvec.url(ntsid, server = "http://ftp2.cits.rncan.gc.ca/pub")
}
\arguments{
\item{ntsid}{A single NTS Reference as generated by nts().}

\item{server}{The server to download from (default: \url{http://ftp2.cits.rncan.gc.ca/pub})}
}
\value{
A URL where the given data can be found.
}
\description{
Get CanVec or CanVec+ data URL based on the NTS Reference (as generated by
\code{nts()}) provided. The URL generated may or may not exist depending whether
or not the sheet or area is available. CanVec data is available by mapsheet
(e.g. 021H01; 1:50k), CanVec+ data is available by map area (e.g. 021H; 1:250k).
If a the \code{ntsid} provided is a 1:50k reference, a CanVec URL will be generated.
Otherwise, a CanVec+ url is generated.
}

