% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/canvec.R
\name{makebbox}
\alias{makebbox}
\title{Create a bounding box like that returned by sp::bbox()
from a more preditable format}
\usage{
makebbox(n, e, s, w)
}
\description{
Create a bounding box like that returned by sp::bbox()
from a more preditable format
}

