% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntstiles.R
\name{nts.bbox}
\alias{nts.bbox}
\title{Get Bounding Box of an NTS Reference}
\usage{
nts.bbox(ntsid)
}
\arguments{
\item{ntsid}{One or more NTS References as generated by \code{nts()}}
}
\value{
A bbox like that returned by \code{sp::bbox()}, or a list of
such objects.
}
\description{
Calculates the bounding box in latitude/longitude described by
a particular NTS Reference.
}
\examples{
nts.bbox(nts('21h'))

}

