% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntstiles.R
\docType{data}
\name{nts.SCALE250K}
\alias{nts.SCALE250K}
\title{A constant denoting NTS Map Area (1:250k) scale (1)
 @export}
\format{An object of class \code{numeric} of length 1.}
\usage{
nts.SCALE250K
}
\description{
A constant denoting NTS Map Area (1:250k) scale (1)
 @export
}
\keyword{datasets}

