% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ntstiles.R
\docType{data}
\name{nts.SCALE50K}
\alias{nts.SCALE50K}
\title{A constant denoting NTS Map Sheet (1:50k) scale (2)}
\format{\preformatted{ num 2
}}
\usage{
nts.SCALE50K
}
\description{
A constant denoting NTS Map Sheet (1:50k) scale (2)
}
\keyword{datasets}

