% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntstiles.R
\docType{data}
\name{nts.SCALESERIES}
\alias{nts.SCALESERIES}
\title{A contstant denoting NTS Series scale (0)}
\format{An object of class \code{numeric} of length 1.}
\usage{
nts.SCALESERIES
}
\description{
A contstant denoting NTS Series scale (0)
}
\keyword{datasets}

